Adding dynamic depth and arc-line drawing.
Dynamic depth additions allow the depth for a given treeGraphInstance to be dynamically determined by looking for the maximum depth that does not exceed a set maximum for the number of named nodes to be displayed (also available are functions for counting all nodes, or ends, so this could be changed later).

Arc-and-line drawing is an alternative drawing style added so that the relative depth of nodes (which is distance from the base node) is more clear. This is especially useful for trees with distances; trees without distances (all nodes are equidistant) this isn't particularly useful.