Speed up zooms & do radial averaging on animation
Zooming animation steps are larger to speed up when multiple nodes are being traverse, in proportion to how many traversals need to be made.

Zooming is now done with radial averages between two tree states, not xy averages. This is how the old tree did it, it is more aesthetically pleasing because branches move outward as well as get smaller while going deeper into the tree.