Adding tree coloring and search-for-node
Default coloring is based on absolute depth in the tree.

Searching has been added allowing user to type in a string and looks for a matching name. One is chosen by clicking on it in the displayed results, and then the tree is colored according to distance from that target. The path of ancestor nodes to the target is also highlighted, and the up/right arrows move forward towards the target.